Better log message if auth-server can't be reached
Provide a meaningful log message if the auth-server cannot be reached to
validate a provided access token.

This commit also turns the connector in the
`AccessTokenValidationService` into a field instead of creating it for
every request.

It adds one test for the code added, however the whole class lacks
tests, we should add them.