Switch order of web and REST security
Anticipate the update to Spring Boot 1.3.1 and therefore Spring
Security 4.0.3.
In the new Spring Security release it's not possible to overwrite a
`requestMatcher()` call on `HttpSecurity` anymore.
This way `NotOAuthRequestMatcher` can't be overridden and the REST
security filter chain acts as the catch all handler.
This makes it the last chain in the list of security filter chains.

The Web security filter chain now takes the last free slot (2).
So the final order of filter chains is now:

- 0: OAuth authorization server
- 1: Facebook client credentials
- 2: Web
- 3: REST (OAuth resource server)