Raise required Java version to Java 8
This commit updates the required java version of java to java 8 and
updates the documentation to reflect that. It also updates the suggested
PostgreSQL version to 9.4 in `detailed-reference-installation.md` as the
update of the docker image to java 8 made a migration to this version
necessary and the documentation should reflect that.