Load configuration and assets from the file system
Add notion of `OSIAM_HOME`, an arbitrary folder in the file system that
contains the static assets and the configuration.

The home directory will default to `$HOME/.osiam`, if not specified.
This seems to be a good place, as the current user will certainly have
access to this folder.
It's also a "best practice" by many other applications.

Use a `ReloadableResourceBundleMessageSource`, because it's the only one
that can load message source bundles from the file system.