Refactor LoginController
`OsiamCachingAuthenticationFailureHandler` sets an LDAP-related error
message in the session.
Passing the message to the view was some kind of magical before: No one
knows how and why it worked.
Explicitly read the error message from the session now and add it to the
model.
Use `ModelAndView` in `/login/error` to simplify error handling.

Extract common model attributes as `@ModelAttribute` annotated methods.
This removes the duplication of setting it.

Model attribute `locked` isn't used by the login template, so remove it.