Squash Flyway migrations
This unifies the first 3 database versions, that reflect the way the
database schema looks like in the latest 2.x version.
The idea is that version 1 will be the new baseline, so that on an
upgrade from 2.x to 3.0, Flyway will mark this schema as applied
(baselined) and won't run it.
After that, version 2 will be applied, that is currently just changing
the scopes of the well-known clients.
Version 2 should be used for all upcoming automatic changes, that deal
with old data, not updated schemas, like removing the auth-server
client, renaming the auth-server extension, etc.

The PostgreSQL scripts have been re-generated by Hibernate's auto DDL
feature and exported with IntelliJ.
It had one unique constraint missing: `osiam_client.id`.
The MySQL scripts are the same as before, but combined.