Make sure core schema is present and valid
The last refactoring of the validation removed the ability to validate
the schema of scim users and groups, i.e. It was possible to provide
users and groups with a wrong or no schema present.

This commit fixes it by providing a custom validator and annotation to
validate the provided schemas.