Initialize home directory via command line
This adds a command to OSIAM, that allows one to initialize the home
directory without starting OSIAM.
The command parser is dead-simple and should be replaced by something
more sophisticated in the future, if we want to support more cli
commands.
The default command is `server` so one don't has to specify it
every time.
Depending on the command, the Application Context will be bootstrapped
with a different `@Configuration` class.
The `InitHome` class represents the simplest (but most sufficient)
configuration to initialize the home directory.
Use a customized `@ComponentScan` definition, to prevent loading the
`InitHome` bean during normal server startup.