Refactor processing of OSIAM's home directory
OSIAM will now load the configuration only from the YAML file in its
home directory. This is done by running the initialization as early as
possible via an `ApplicationListener`. Initialization will take place
when the environment is ready to use, but before external configuration
will be loaded by the `ConfigFileApplicationListener`.

At this early stage of application startup, logging has not yet been set
up. The first time logging is available, is when the
`ApplicationPreparedEvent` is received. To not lose the messages, they
will be logged then. To log whether the home directory was actually
initialized, the flag `hasInitializedHome` has been introduced.

Also introduce a `PropertySource`, that will hold a checked and munged
value of the user-supplied home directory. This is because it could be
empty or a relative path, which does not work with the upcoming support
of a file-based H2 database.