Initialize database via command line
This adds a command to OSIAM, that allows one to initialize the home
directory without starting OSIAM. The home directory must already be
initialized and any configuration has to be done, before this command
can be used. This should prevent accidental faulty database migrations,
because the command fails, if no configuration can be found in the home
directory. Initialization will be switched off in `OsiamHome` to realize
this.

Resolves #167