Remove migration of client scopes
In the latest version of the add-ons we've already dropped support for
the method-based scopes. This way, users had to do the migration anyway
by now. As we will only support a migration from the latest version of
OSIAM and all it's components, remove the Flyway migration.

Furthermore we don't know how users utilize the "example-client", so
removing the scopes and adding both "ADMIN" and "ME" could break their
setup or even make their OSIAM insecure.

Resolves #166