Make extensions configurable
Read extensions from `osiam.yaml` and create them on start up. If the
extension already exists, don't create it. Add new value classes for
extension and extension field, because the `ExtensionEntity` stores the
fields as `Set` and Spring is unable to bind a YAML list to a `Set`.
This is due to lists are ordered, but sets not. Use Bean validation to
validate the configured extensions, which keeps the code simpler.