Update Spring Boot to version 1.4.1
Remove `@Lob` and change `@Type` to `text` for text fields with variable
length. Hibernate 5 removed the type `StringClobType` and `text` is the
best replacement, as it does not map Strings to CLOBs, which does not
work with PotgreSQL.

Explicitly name some columns, as the naming strategy changed again in
Hibernate 5 and changing the column names in the database schema is
rather hard to do. This commit's scope is to only update Spring Boot.

Remove or replace deprecated things like
`spring.jpa.hibernate.naming_strategy` configuration property or
`org.springframework.boot.context.web.SpringBootServletInitializer`.